[Move] DI from `src/code/common` to `src/code/platform`
