+++ src/base/common/logger.ts
@@ -1,7 +1,7 @@
 import { Disposable } from "src/base/common/dispose";
 import { Emitter, Register } from "src/base/common/event";
 import { mockType } from "src/base/common/util/type";
-import { createDecorator } from "src/code/common/service/instantiationService/decorator";
+import { createDecorator } from "src/code/platform/instantiation/common/decorator";
 
 export const ILogService = createDecorator<ILogService>('log-service');
 export const DEFAULT_LOG_LEVEL = LogLevel.INFO;
