+++ src/code/browser/service/componentService.ts
@@ -1,4 +1,4 @@
-import { createDecorator } from "src/code/common/service/instantiationService/decorator";
+import { createDecorator } from "src/code/platform/instantiation/common/decorator";
 import { IComponent } from "src/code/browser/workbench/component";
 
 export const IComponentService = createDecorator<IComponentService>('component-service');
