+++ src/code/browser/service/keyboardScreenCastService/keyboardScreenCastService.ts
@@ -4,7 +4,7 @@ import { DomUtility } from "src/base/common/dom";
 import { IStandardKeyboardEvent, Keyboard } from "src/base/common/keyboard";
 import { IntervalTimer } from "src/base/common/util/timer";
 import { IKeyboardService } from "src/code/browser/service/keyboardService";
-import { createDecorator } from "src/code/common/service/instantiationService/decorator";
+import { createDecorator } from "src/code/platform/instantiation/common/decorator";
 
 export const IKeyboardScreenCastService = createDecorator<IKeyboardScreenCastService>('keyboard-screencast-service');
