+++ src/code/browser/workbench/actionBar/actionBar.ts
@@ -1,14 +1,14 @@
 import { Component, ComponentType, IComponent } from 'src/code/browser/workbench/component';
 import { ContextMenuType, Coordinate } from 'src/base/browser/secondary/contextMenu/contextMenu';
-import { createDecorator } from 'src/code/common/service/instantiationService/decorator';
+import { createDecorator } from 'src/code/platform/instantiation/common/decorator';
 import { IContextMenuService } from 'src/code/browser/service/contextMenuService';
 import { IComponentService } from 'src/code/browser/service/componentService';
 import { ActionButton } from 'src/code/browser/workbench/actionBar/actionButton';
 import { WidgetBar } from 'src/base/browser/secondary/widgetBar/widgetBar';
 import { Orientation } from 'src/base/common/dom';
 import { Icons } from 'src/base/browser/icon/icons';
-import { registerSingleton } from 'src/code/common/service/instantiationService/serviceCollection';
-import { ServiceDescriptor } from 'src/code/common/service/instantiationService/descriptor';
+import { registerSingleton } from 'src/code/platform/instantiation/common/serviceCollection';
+import { ServiceDescriptor } from 'src/code/platform/instantiation/common/descriptor';
 import { Emitter, Register } from 'src/base/common/event';
 
 export const IActionBarService = createDecorator<IActionBarService>('action-bar-service');
