+++ src/code/browser/workbench/workspace/titleBar/titleBar.ts
@@ -2,7 +2,7 @@ import { Component } from 'src/code/browser/workbench/component';
 import { WindowBarComponent } from 'src/code/browser/workbench/workspace/titleBar/windowBar';
 import { WorkspaceComponentType } from 'src/code/browser/workbench/workspace/workspace';
 import { IComponentService } from 'src/code/browser/service/componentService';
-import { IInstantiationService } from 'src/code/common/service/instantiationService/instantiation';
+import { IInstantiationService } from 'src/code/platform/instantiation/common/instantiation';
 
 export const enum TitleBarComponentType {
     functionBar = 'function-bar',
