+++ src/code/electron/main.ts
@@ -8,8 +8,8 @@ import { Strings } from 'src/base/common/util/string';
 import { DiskFileSystemProvider } from 'src/code/platform/files/node/diskFileSystemProvider';
 import { GlobalConfigService, IGlobalConfigService, IUserConfigService, UserConfigService } from 'src/code/platform/configuration/electron/configService';
 import { FileService, IFileService } from 'src/code/platform/files/common/fileService';
-import { IInstantiationService, InstantiationService } from 'src/code/common/service/instantiationService/instantiation';
-import { ServiceCollection } from 'src/code/common/service/instantiationService/serviceCollection';
+import { IInstantiationService, InstantiationService } from 'src/code/platform/instantiation/common/instantiation';
+import { ServiceCollection } from 'src/code/platform/instantiation/common/serviceCollection';
 import { ILoggerService } from 'src/code/platform/logger/common/abstractLoggerService';
 import { ConsoleLogger } from 'src/code/platform/logger/common/consoleLoggerService';
 import { FileLoggerService } from 'src/code/platform/logger/common/fileLoggerService';
