+++ src/code/electron/nota.ts
@@ -6,9 +6,9 @@ import { ILogService } from "src/base/common/logger";
 import { getUUID, UUID } from "src/base/node/uuid";
 import { IGlobalConfigService, IUserConfigService } from "src/code/platform/configuration/electron/configService";
 import { IFileService } from "src/code/platform/files/common/fileService";
-import { ServiceDescriptor } from "src/code/common/service/instantiationService/descriptor";
-import { IInstantiationService, IServiceProvider } from "src/code/common/service/instantiationService/instantiation";
-import { ServiceCollection } from "src/code/common/service/instantiationService/serviceCollection";
+import { ServiceDescriptor } from "src/code/platform/instantiation/common/descriptor";
+import { IInstantiationService, IServiceProvider } from "src/code/platform/instantiation/common/instantiation";
+import { ServiceCollection } from "src/code/platform/instantiation/common/serviceCollection";
 import { IEnvironmentService, IMainEnvironmentService } from "src/code/platform/environment/common/environment";
 import { IpcServer } from "src/code/platform/ipc/browser/ipc";
 import { IpcChannel } from "src/code/platform/ipc/common/channel";
