+++ src/code/platform/i18n/i18n.ts
@@ -4,7 +4,7 @@ import * as path from "src/base/common/file/path";
 import { IFileService } from "src/code/platform/files/common/fileService";
 import { isArray, isObject } from "src/base/common/util/type";
 import { Section } from "src/code/platform/section";
-import { createDecorator } from "src/code/common/service/instantiationService/decorator";
+import { createDecorator } from "src/code/platform/instantiation/common/decorator";
 
 export const Ii18nService = createDecorator<Ii18nService>('i18n-service');
