+++ src/code/platform/window/electron/mainWindowService.ts
@@ -4,8 +4,8 @@ import { ILogService } from "src/base/common/logger";
 import { Mutable } from "src/base/common/util/type";
 import { UUID } from "src/base/node/uuid";
 import { IFileService } from "src/code/platform/files/common/fileService";
-import { createDecorator } from "src/code/common/service/instantiationService/decorator";
-import { IInstantiationService } from "src/code/common/service/instantiationService/instantiation";
+import { createDecorator } from "src/code/platform/instantiation/common/decorator";
+import { IInstantiationService } from "src/code/platform/instantiation/common/instantiation";
 import { IEnvironmentService, IMainEnvironmentService } from "src/code/platform/environment/common/environment";
 import { IMainLifeCycleService } from "src/code/platform/lifeCycle/electron/mainLifeCycleService";
 import { ICreateWindowConfiguration, IOpenWindowOpts, IWindowInstance } from "src/code/platform/window/common/window";
